# frontend/openapi.json
{"openapi":"3.0.0","paths":{"/v1":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/v2":{"get":{"operationId":"AppController_getHello2","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/v3":{"get":{"operationId":"AppController_getHello3","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"number"}}}}}}},"/user":{"post":{"operationId":"UserController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}}},"get":{"operationId":"UserController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}}}}},"/user/{id}":{"get":{"operationId":"UserController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Test"}}}}}},"patch":{"operationId":"UserController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}}},"delete":{"operationId":"UserController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}}}}},"info":{"title":"Cats example","description":"The cats API description","version":"999.0","contact":{}},"tags":[{"name":"cats","description":""}],"servers":[],"components":{"schemas":{"CreateUserDto":{"type":"object","properties":{"firstName":{"type":"string","minLength":6},"lastName":{"type":"string"}},"required":["firstName","lastName"]},"User":{"type":"object","properties":{"id":{"type":"number"},"firstName":{"type":"string","minLength":6},"lastName":{"type":"string"},"isEnabled":{"type":"boolean","default":true}},"required":["id","firstName","lastName"]},"Test":{"type":"object","properties":{"firstName":{"type":"string"}},"required":["firstName"]},"UpdateUserDto":{"type":"object","properties":{}}}}}